Remove/merge the DTD/Editor/Sidebar classes used to support both protocols
No functional changes here, just merging some classes and removing anything that was redundant. Previously a lot of this code was split across classes to provide the functionality across both postMessage and DTD, but now we only support DTD.
