+++ packages/devtools_app/lib/src/service/editor/editor_client.dart
@@ -10,96 +10,20 @@ import 'package:dtd/dtd.dart';
 import '../../shared/analytics/constants.dart';
 import 'api_classes.dart';
 
-/// An interface to services provided by an editor.
+/// A client wrapper that connects to an editor over DTD.
 ///
 /// Changes made to the editor services/events should be considered carefully to
 /// ensure they are not breaking changes to already-shipped editors.
-abstract class EditorClient extends DisposableController {
-  Future<void> close();
+class EditorClient extends DisposableController
+    with AutoDisposeControllerMixin {
 
-  /// The ID to use for analytics events.
-  String get gaId;
-
-  /// Whether the connected editor supports the `getDevices` method.
-  bool get supportsGetDevices;
-
-  /// Whether the connected editor supports the `getDebugSessions` method.
-  bool get supportsGetDebugSessions;
-
-  /// Whether the connected editor supports the `selectDevice` method.
-  bool get supportsSelectDevice;
-
-  /// Whether the connected editor supports the `hotReload` method.
-  bool get supportsHotReload;
-
-  /// Whether the connected editor supports the `hotRestart` method.
-  bool get supportsHotRestart;
-
-  /// Whether the connected editor supports the `openDevToolsPage` method.
-  bool get supportsOpenDevToolsPage;
-
-  /// Whether the connected editor supports the `forceExternal` flag in the
-  /// params for `openDevToolsPage`.
-  bool get supportsOpenDevToolsForceExternal;
-
-  /// A stream of [EditorEvent]s from the editor.
-  Stream<EditorEvent> get event;
-
-  /// A stream of events of when editor service methods/capabilities change.
-  Stream<ServiceRegistrationChange> get editorServiceChanged;
-
-  /// Gets the set of currently available devices from the editor.
-  Future<GetDevicesResult> getDevices();
-
-  /// Gets the set of currently active debug sessions from the editor.
-  Future<GetDebugSessionsResult> getDebugSessions();
-
-  /// Requests the editor selects a specific device.
-  ///
-  /// It should not be assumed that calling this method succeeds (if it does, a
-  /// `deviceSelected` event will provide the appropriate update).
-  Future<void> selectDevice(EditorDevice? device);
-
-  /// Requests the editor Hot Reloads the given debug session.
-  Future<void> hotReload(String debugSessionId);
-
-  /// Requests the editor Hot Restarts the given debug session.
-  Future<void> hotRestart(String debugSessionId);
-
-  /// Requests the editor opens a DevTools page for the given debug session.
-  Future<void> openDevToolsPage(
-    String? debugSessionId, {
-    String? page,
-    bool? forceExternal,
-    bool? requiresDebugSession,
-    bool? prefersDebugSession,
-  });
-
-  /// Requests the editor enables a new platform (for example by running
-  /// `flutter create` to add the native project files).
-  ///
-  /// This action may prompt the user so it should not be assumed that calling
-  /// this method succeeds (if it does, a `deviceChanged` event will provide
-  /// the appropriate updates).
-  Future<void> enablePlatformType(String platformType);
-}
-
-/// An implementation of [EditorClient] that connects to an editor over DTD.
-///
-/// Changes made to the editor services/events should be considered carefully to
-/// ensure they are not breaking changes to already-shipped editors.
-class DtdEditorClient extends EditorClient with AutoDisposeControllerMixin {
-  // TODO(dantup): Merge this into EditorClient once the postMessage version
-  //  is removed.
-
-  DtdEditorClient(this._dtd) {
+  EditorClient(this._dtd) {
     unawaited(initialized); // Trigger async initialization.
   }
 
   final DartToolingDaemon _dtd;
   late final initialized = _initialize();
 
-  @override
   String get gaId => EditorSidebar.id;
 
   Future<void> _initialize() async {
@@ -191,59 +115,47 @@ class DtdEditorClient extends EditorClient with AutoDisposeControllerMixin {
   }
 
   /// Close the connection to DTD.
-  @override
   Future<void> close() => _dtd.close();
 
-  @override
   bool get supportsGetDevices => _supportsGetDevices;
   var _supportsGetDevices = false;
 
-  @override
   bool get supportsGetDebugSessions => _supportsGetDebugSessions;
   var _supportsGetDebugSessions = false;
 
-  @override
   bool get supportsSelectDevice => _supportsSelectDevice;
   var _supportsSelectDevice = false;
 
-  @override
   bool get supportsHotReload => _supportsHotReload;
   var _supportsHotReload = false;
 
-  @override
   bool get supportsHotRestart => _supportsHotRestart;
   var _supportsHotRestart = false;
 
-  @override
   bool get supportsOpenDevToolsPage => _supportsOpenDevToolsPage;
   var _supportsOpenDevToolsPage = false;
 
-  @override
   bool get supportsOpenDevToolsForceExternal =>
       _supportsOpenDevToolsForceExternal;
   var _supportsOpenDevToolsForceExternal = false;
 
   /// A stream of [EditorEvent]s from the editor.
-  @override
   Stream<EditorEvent> get event => _eventController.stream;
   final _eventController = StreamController<EditorEvent>();
 
   /// A stream of events of when editor services are registrered or
   /// unregistered.
-  @override
   Stream<ServiceRegistrationChange> get editorServiceChanged =>
       _editorServiceChangedController.stream;
   final _editorServiceChangedController =
       StreamController<ServiceRegistrationChange>();
 
-  @override
   Future<GetDevicesResult> getDevices() async {
     final response = await _call(EditorMethod.getDevices);
     return GetDevicesResult.fromJson(response.result);
   }
 
   /// Gets the set of currently active debug sessions from the editor.
-  @override
   Future<GetDebugSessionsResult> getDebugSessions() async {
     final response = await _call(EditorMethod.getDebugSessions);
     return GetDebugSessionsResult.fromJson(response.result);
@@ -253,7 +165,6 @@ class DtdEditorClient extends EditorClient with AutoDisposeControllerMixin {
   ///
   /// It should not be assumed that calling this method succeeds (if it does, a
   /// `deviceSelected` event will provide the appropriate update).
-  @override
   Future<void> selectDevice(EditorDevice? device) async {
     await _call(
       EditorMethod.selectDevice,
@@ -261,7 +172,6 @@ class DtdEditorClient extends EditorClient with AutoDisposeControllerMixin {
     );
   }
 
-  @override
   Future<void> hotReload(String debugSessionId) async {
     await _call(
       EditorMethod.hotReload,
@@ -269,7 +179,6 @@ class DtdEditorClient extends EditorClient with AutoDisposeControllerMixin {
     );
   }
 
-  @override
   Future<void> hotRestart(String debugSessionId) async {
     await _call(
       EditorMethod.hotRestart,
@@ -277,7 +186,6 @@ class DtdEditorClient extends EditorClient with AutoDisposeControllerMixin {
     );
   }
 
-  @override
   Future<void> openDevToolsPage(
     String? debugSessionId, {
     String? page,
@@ -297,7 +205,6 @@ class DtdEditorClient extends EditorClient with AutoDisposeControllerMixin {
     );
   }
 
-  @override
   Future<void> enablePlatformType(String platformType) async {
     await _call(
       EditorMethod.enablePlatformType,
